pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "config-validator"
version = "0.1.0"
description = "Validate YAML configuration files with rules + plugin architecture"
 
 
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Maryam Golozar", email = "maryam.golozar@gmail.com"}
]

dependencies = [
"PyYAML>=6.0",
]


[project.optional-dependencies]
dev = [
"pytest>=8.0",
"pytest-cov>=5.0",
"ruff>=0.6",
"mypy>=1.10",
]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]


[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
ignore_missing_imports = true


[project.scripts]
config-validator = "config_validator.cli:main"